requirements.txt
fastapi
uvicorn
requests
faiss-cpu
sentence-transformers
groq
python-multipart
